# backend/secret_santa/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Wichtel API
  version: 1.0.0
servers:
- url: http://localhost:8080/api/v1
  description: Dev server
tags:
- name: wichtel
  description: Wichtel client api
paths:
  /settings:
    get:
      tags:
      - wichtel
      operationId: get_settings
      x-openapi-router-controller: secret_santa.controllers.wichtel_controller
      responses:
        "200": 
          description: ok
  /assignments/{participantId}:
    get:
      tags:
      - wichtel
      summary: Returns the assignments of the participant
      operationId: get_draw
      parameters:
      - name: participantId
        in: path
        description: ID of participant
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Participant status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetails'
        "404":
          description: no assignments for participant
          content: {}
      x-openapi-router-controller: secret_santa.controllers.wichtel_controller              
  /participations/{participantId}:
    get:
      tags:
      - wichtel
      summary: Returns the participation status
      operationId: get_participation
      parameters:
      - name: participantId
        in: path
        description: ID of participant
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Participant status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/participation'
        "404":
          description: Participant not found
          content: {}
      x-openapi-router-controller: secret_santa.controllers.wichtel_controller          
  /participations/{participantId}:
    get:
      tags:
      - wichtel
      summary: Returns the participation status
      operationId: get_participation
      parameters:
      - name: participantId
        in: path
        description: ID of participant
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Participant status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/participation'
        "404":
          description: Participant not found
          content: {}
      x-openapi-router-controller: secret_santa.controllers.wichtel_controller
    put:
      tags:
      - wichtel
      summary: Update a participation
      operationId: update_participation
      parameters:
      - name: participantId
        in: path
        description: ID of participant
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        description: participation details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/participation'
        required: true
      responses:
        "204":
          description: Participant updated
          content: {}
        "400":
          description: Invalid request
          content: {}
        "404":
          description: Participant not found
          content: {}
      x-openapi-router-controller: secret_santa.controllers.wichtel_controller
  /users/{participantId}:
    get:
      tags:
      - wichtel
      summary: Returns the participant details
      operationId: get_participant_details
      parameters:
      - name: participantId
        in: path
        description: ID of participant
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Participant status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetails'
        "404":
          description: Participant not found
          content: {}
      x-openapi-router-controller: secret_santa.controllers.wichtel_controller
components:
  schemas:
    participation:
      type: object
      properties:
        participating:
          type: boolean
      example:
        participating: true
    PersonDetails:
      type: object
      properties:
        firstName:
          type: string
          example: Samuel
        lastName:
          type: string
          example: Vennmann
      example:
        firstName: Samuel
        lastName: Vennmann